+++ client/src/main/java/org/sportApp/training/TrainingEventDto.java
@@ -3,10 +3,10 @@ package org.sportApp.training;
 import java.time.LocalDate;
 
 public class TrainingEventDto {
-    private int id;
+    private long id;
     private LocalDate date;
     private boolean completed;
-    private int userId;
+    private long userId;
     private String comment;
     private TrainingDto trainingDto;
 
@@ -19,14 +19,14 @@ public class TrainingEventDto {
     public void setDate(LocalDate selectedDate) {
     }
 
-    public void setId(int i) {
+    public void setId(long i) {
         id = i;
     }
 
     public void setCompleted(boolean b) {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
